requirements.txt
certifi==2018.11.29
chardet==3.0.4
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
pkg-resources==0.0.0
requests==2.21.0
smmap2==2.0.5
urllib3==1.24.1
